src/agent/requirements/requirements.txt
# azure-ai-projects @ git+https://github.com/Azure/azure-sdk-for-python.git@main#egg=azure-ai-projects&subdirectory=sdk/ai/azure-ai-projects #test
# promptflow-core==1.17.2
# promptflow-devkit==1.17.2
# promptflow-tracing==1.17.2
# pydantic==2.10.6
#pywin32-ctypes==0.2.3
#pywin32==310
#pywinpty==2.0.15
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.0
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
asyncio==3.4.3
attrs==25.3.0
azure-ai-contentsafety==1.0.0
azure-ai-evaluation==1.3.0
azure-ai-inference==1.0.0b6
azure-ai-projects==1.0.0b8
azure-common==1.1.28
azure-core-tracing-opentelemetry==1.0.0b11
azure-core==1.32.0
azure-identity==1.21.0
azure-kusto-data==5.0.2
azure-mgmt-authorization==4.0.0
azure-mgmt-core==1.5.0
azure-mgmt-resource==23.3.0
azure-mgmt-resourcegraph==8.0.0
azure-monitor-opentelemetry-exporter==1.0.0b35
azure-monitor-opentelemetry==1.6.5
azure-search-documents==11.6.0b7
azure-storage-blob==12.25.0
babel==2.17.0
backrefs==5.8
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
bs4==0.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
csscompressor==0.9.5
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.16
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
fixedint==0.1.6
flask-cors==5.0.1
flask-restx==1.3.0
Flask==3.1.0
fqdn==1.5.1
ghp-import==2.1.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
hjson==3.1.0
htmlmin2==0.1.13
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ijson==3.3.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython_pygments_lexers==1.1.1
ipython==9.0.2
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsmin==3.0.1
json5==0.10.0
jsonpointer==3.0.0
jsonschema-specifications==2024.10.1
jsonschema==4.23.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server_terminals==0.5.3
jupyter_server==2.15.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter==1.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
jupyterlab==4.3.6
jupytext==1.16.7
keyring==24.3.1
lxml==5.3.1
markdown-it-py==3.0.0
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.1.3
mkdocs-get-deps==0.2.0
mkdocs-git-committers-plugin-2==2.5.0
mkdocs-git-revision-date-localized-plugin==1.4.5
mkdocs-git-revision-date-plugin==0.3.2
mkdocs-glightbox==0.4.0
mkdocs-jupyter==0.25.1
mkdocs-macros-plugin==1.3.7
mkdocs-material-extensions==1.3.1
mkdocs-material==9.6.9
mkdocs-minify-plugin==0.8.0
mkdocs-open-in-new-tab==1.0.8
mkdocs-render-swagger-plugin==0.1.2
mkdocs-table-reader-plugin==3.1.0
mkdocs-video==1.5.0
mkdocs==1.6.1
modelcontextprotocol
more-itertools==10.6.0
msal-extensions==1.3.1
msal==1.32.0
msrest==0.7.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook_shim==0.2.4
notebook==7.3.3
numpy==2.2.4
oauthlib==3.2.2
openai==1.68.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-dbapi==0.52b1
opentelemetry-instrumentation-django==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-langchain==0.39.0
opentelemetry-instrumentation-openai==0.39.0
opentelemetry-instrumentation-psycopg2==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-urllib==0.52b1
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-wsgi==0.52b1
opentelemetry-instrumentation==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
overrides==7.7.0
packaging==24.2
paginate==0.5.7
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pillow==10.3.0 #was 11.1.0
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==6.1.1
pure_eval==0.2.3
pycparser==2.22
pydantic
pydantic_core==2.27.2
pydantic<2.11,!=2.10.0,!=2.10.1,!=2.10.2,!=2.10.3
pydash==7.0.7
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.14.3
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2025.1
pyyaml_env_tag==0.1
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests-oauthlib==2.0.0
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
ruamel.yaml.clib==0.2.12
ruamel.yaml==0.18.10
semantic-kernel[azure]==1.26.1
Send2Trash==1.8.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.46.1
strictyaml==1.7.3
super_collections==0.5.3
tabulate==0.9.0
tavily-python==0.5.1
#tavily==0.7.0
termcolor==2.5.0
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
typing==3.7.4.3
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn[standard]==0.27.1
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets<15,>=13
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
zipp==3.21.0